api basic functions in db with testing

+++ api/app/routes/driver.py
@@ -29,21 +29,25 @@ def read_trip_info(
     user: Annotated[UserIn, Depends(get_current_user)],
     db: Annotated[Session,Depends(get_db)],
 ):
-    ret = db.get_data_trips(user.user_id)
-    trip_set = set([trip['trip_id'] for trip in ret])
+    ret = db.get_data_trips_by_userid(user.user_id)
     trips = []
-    for trip_id in trip_set:
-        t = dict()
-        t['path'] = []
-        for trip in ret:
-            if trip_id != trip['trip_id']:
-                continue
-            t['trip_id'] = trip_id
-            t["driver_name"] = trip['driver_name']
-            t['available_seats'] = trip['available_seats']
-            boarding_time = datetime.strftime(trip['boarding_time'],'%Y-%m-%d %H:%M:%S')
-            t['path'].append({'location':trip['path'],'time':boarding_time})
-        trips.append(t)
+    for trip in ret:
+        print(trip)
+        single_trip = {
+            'trip_id' : trip['trip_id'],
+            'driver_name' : user.username,
+            'available_seats' : trip['available_seats'],
+            'departure' : {
+                'location':trip['departure'],
+                'time':datetime.strftime(trip['boarding_time'],'%Y-%m-%d %H:%M:%S'),
+            },
+            'destination' : {
+                'location':trip['destination'],
+                'time':datetime.strftime(trip['boarding_time'],'%Y-%m-%d %H:%M:%S'),
+            }
+        }
+        trips.append(single_trip)
+    return sorted(trips,key=lambda x:x['trip_id'])
     return trips
 
 class ReservedLocation(BaseModel):
@@ -68,19 +72,18 @@ def read_reserved_trip_info(
     if isinstance(locations,str):
         return Response(status_code=status.HTTP_404_NOT_FOUND)
     passengers = db.get_data_passengers(trip_id)
-    print(f'-----------------------------')
-    print(passengers)
-    print(f'-----------------------------')
-
-    print(f'-----------------------------')
-
-    print(locations)
-    print(f'-----------------------------')
 
     # name : str , time : str , trip_id
     info = []
     for location in locations:
-        dic = {"point":{"location":location["name"],"time":datetime.strftime(location["time"],'%Y-%m-%d %H:%M:%S')},"boarding":[],"Alighting":[]}
+        dic = {
+            'point':  {
+                'location' : location['name'],
+                'time' : datetime.strftime(location["time"],'%Y-%m-%d %H:%M:%S')
+            },
+            'boarding' : [],
+            'Alighting' : []
+        }
         if isinstance(passengers,str):
             info.append(dic)
             continue
@@ -88,30 +91,10 @@ def read_reserved_trip_info(
             data = db.get_data_users_byid(passenger['user_id'])
             if passenger['departure'] == location['name']:
                 dic["boarding"].append(data['username'])
-            if passenger['destionation'] == location['name']:
+            if passenger['destination'] == location['name']:
                 dic["Alighting"].append(data['username'])
         info.append(dic)
-    print(info)
     return info
-
-
-    # info = {
-    #     "point": {
-    #         "location" : "NYCU",
-    #         "time" : "2023/6/2 17:00"
-    #     },
-    #     "boarding":[
-    #         "AAA",
-    #         "BBB",
-    #         "CCC",
-    #     ],
-    #     "Alighting":[
-    #         "AAA",
-    #         "BBB",
-    #         "CCC",
-    #     ]
-    # }
-    # return [info,info]
 
 
 @router.post(
@@ -130,10 +113,11 @@ def new_trip(
 ):
     #2011-11-04T00:05:23
     query.boarding_time = datetime.fromisoformat(query.boarding_time)
-    req = {}
-    req['user_id'] = user.user_id
-    req['available_seats'] = query.available_seats
-    req['boarding_time'] = query.boarding_time
+    req = {
+        'user_id' : user.user_id,
+        'available_seats' : query.available_seats,
+        'boarding_time' : query.boarding_time
+    }
     res = db.insert_data_trips(req)
     trip_id = db.get_data_tripid(req)
     if res != "SUCCESS":
@@ -141,14 +125,18 @@ def new_trip(
     if query.path == []:
         return Response(status_code=status.HTTP_406_NOT_ACCEPTABLE)
     for p in query.path:
-        req = {}
-        req['trip_id'] = trip_id['trip_id']
-        req['name'] = p
-        req['time'] = query.boarding_time
+        req = {
+            'trip_id' : trip_id['trip_id'],
+            'name' : p,
+            'time' : query.boarding_time
+        }
         db.insert_data_locations(req)
     first_location = db.get_data_locationsid(trip_id['trip_id'],query.path[0]) 
     last_location = db.get_data_locationsid(trip_id['trip_id'],query.path[-1])
-    dic = {"departure":first_location["location_id"],"destination":last_location["location_id"]}
+    dic = {
+        'departure' : first_location['location_id'],
+        'destination' : last_location['location_id'],
+    }
     db.update_data_trips(trip_id['trip_id'],dic)
 
     return Response(status_code=status.HTTP_201_CREATED)
@@ -159,6 +147,7 @@ def new_trip(
     responses={
         status.HTTP_200_OK: {"content": None},
         status.HTTP_404_NOT_FOUND: {"description": "There is no such trip."},
+        status.HTTP_403_FORBIDDEN: {"description": "passenger or location remove error"}
     }
 )
 def remove_trip(
@@ -166,6 +155,10 @@ def remove_trip(
     _: Annotated[UserIn, Depends(get_current_user)],
     db: Annotated[Session,Depends(get_db)],
 ):
+
+    db.delete_data_passengers(trip_id)
+    if db.delete_data_locations(trip_id) != "SUCCESS":
+        return Response(status_code=status.HTTP_403_FORBIDDEN)
     if db.delete_data_trips(trip_id) != "SUCCESS":
         return Response(status_code=status.HTTP_404_NOT_FOUND)
